+++ run-ui.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Activate the virtual environment
+source venv/bin/activate
+
+# Run the web UI
+python -m src.hassy_normalizer.web_ui.server